change: make opencv as optional build tags
