+++ hrp/internal/uixt/gesture.go
@@ -1,8 +1,9 @@
+//go:build opencv
+
 package uixt
 
 import (
 	"image"
-	"sort"
 
 	"github.com/electricbubble/gwda"
 )
