+++ scripts/build.sh
@@ -15,6 +15,7 @@ mkdir -p "output"
 bin_path="output/hrp"
 
 # build
+# optional build tags: opencv
 go build -ldflags '-s -w' -o "$bin_path" hrp/cmd/cli/main.go
 
 # check output and version
